🚚 Moving all methodology ressources
