+++ pages/methodologie/dailyPositives-hospitalisations.vue
@@ -20,10 +20,10 @@
 import {
   CHARTS_DEFAULT_OPTIONS_AXES_TIME,
   CHARTS_DEFAULT_OPTIONS_AXES_Y,
-} from '~/constants'
-import mixinCharts from '~/mixins/mixin-charts'
-import mixinDatasets from '~/mixins/mixin-datasets'
-import mixinDates from '~/mixins/mixin-dates'
+} from '@/constants'
+import mixinCharts from '@/mixins/mixin-charts'
+import mixinDatasets from '@/mixins/mixin-datasets'
+import mixinDates from '@/mixins/mixin-dates'
 
 export default {
   name: 'PageMethodologie',
@@ -50,8 +50,8 @@ export default {
     }
   },
   mounted() {
-    const hospitalizationsResp = require('~/ressources/hospitalizations.json')
-    const dailyPositivesRest = require('~/ressources/dailyPositives.json').data
+    const hospitalizationsResp = require('@/ressources/hospitalizations.json')
+    const dailyPositivesRest = require('@/ressources/dailyPositives.json').data
       .all
 
     const dailyPositivesData = dailyPositivesRest.map((value, index, array) => {
